refactor. htmlcov added
