+++ tests/test_burger.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 import pytest
-from core.burger import Burger
+from praktikum.burger import Burger
 
 
 class TestBurger:
