+++ tests/test_ingredient.py
@@ -1,5 +1,5 @@
 import pytest
-from core.ingredient import Ingredient
+from praktikum.ingredient import Ingredient
 
 
 class TestIngredient:
